follow XDG data dir layout
JFileSync3 is not relocatable, not because of
requirements of the JAR itself, or its wrapping shell
script, but because the desktop file hard-codes a path
to the shell script, and to the icon.

The Freedesktop.org data directory layout allows for
relative paths in the desktop file on Freedesktop.org
conforming systems, and thus makes it relocatable.

Link: https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html
Link: https://specifications.freedesktop.org/menu-spec/latest/paths.html
Link: https://specifications.freedesktop.org/icon-theme-spec/latest/#icon_lookup

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>