use valid r-DNS application ID
The Freedesktop.org desktop entry specification
recommends an ID "controlled by the author", while the
AppStream specification makes it a requirement.

The desktop entry uses JFileSync3, and the project
internally jfs as inherited from JFileSync. Both are not
registered TLDs.

io.github.mgoellnitz.JFileSync3 is used as it is
guaranteed to be a valid ID through existence of this
repo.

Link: https://specifications.freedesktop.org/desktop-entry-spec/latest/file-naming.html
Link: https://www.freedesktop.org/software/appstream/docs/chap-Metadata.html

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>